requirements.txt
panda
aiohttp
asyncio
humanize
boto3
botocore
python-dotenv